clean up scrutinizer issues

+++ src/Font/Backend/File/Table/HHeaTable.php
@@ -26,11 +26,11 @@ use PdfGenerator\Font\Backend\File\TableVisitor;
 class HHeaTable extends BaseTable
 {
     /**
-     * number of tables contained.
+     * version of table.
      *
-     * @ttf-type uint16
+     * @ttf-type fixed
      *
-     * @var int
+     * @var float
      */
     private $version;
 
@@ -128,9 +128,9 @@ class HHeaTable extends BaseTable
      * caret offset
      * by which amount the highlight of the slanted character should be shifted.
      *
-     * @ttf-type fword
+     * @ttf-type int16
      *
-     * @var float
+     * @var int
      */
     private $caretOffset;
 
@@ -153,12 +153,12 @@ class HHeaTable extends BaseTable
      */
     private $numOfLongHorMetrics;
 
-    public function getVersion(): int
+    public function getVersion(): float
     {
         return $this->version;
     }
 
-    public function setVersion(int $version): void
+    public function setVersion(float $version): void
     {
         $this->version = $version;
     }
@@ -253,12 +253,12 @@ class HHeaTable extends BaseTable
         $this->caretSlopeRun = $caretSlopeRun;
     }
 
-    public function getCaretOffset(): float
+    public function getCaretOffset(): int
     {
         return $this->caretOffset;
     }
 
-    public function setCaretOffset(float $caretOffset): void
+    public function setCaretOffset(int $caretOffset): void
     {
         $this->caretOffset = $caretOffset;
     }
